audit-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
psycopg2-binary
pika
python-dotenv
pydantic-settings
python-jose[cryptography]
